backend/requirements.txt
fastapi==0.115.0
pydantic-settings==2.5.2
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pgvector==0.3.2
python-multipart==0.0.12
pillow==10.4.0
google-genai==0.8.0
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.2
numpy==1.26.4
boto3==1.35.36
